cleanup in prep to release to mvn central

+++ src/main/java/colt/nicity/core/lang/MinMaxInt.java
@@ -141,7 +141,7 @@ public class MinMaxInt {
      * @return
      */
     public double mean() {
-        return sum / (int) count;
+        return sum / count;
     }
     /**
      *
